Extract LoadStatement extrator (#306)
