+++ src/main/scala/me/rexim/morganey/autocompletion/ReplAutocompletion.scala
@@ -10,6 +10,8 @@ import me.rexim.morganey.syntax.LambdaParser
 import me.rexim.morganey.syntax.Language.identifier
 import me.rexim.morganey.util._
 
+import extractors.LoadStatement
+
 object ReplAutocompletion {
 
   // TODO(#257): use classpath modules in ReplAutocompletion
@@ -82,32 +84,6 @@ object ReplAutocompletion {
       case (Nil, false)      => moduleFinder.topLevelDefinitions().map(moduleName)
       // load math.ari|
       case (xs :+ x, false)  => everythingIn(xs, matches(_, x))
-    }
-  }
-
-  private object LoadStatement {
-    val zeroLoad = (Nil, false)
-
-    def pathInformation(path: String) =
-      if (path.isEmpty) {
-        zeroLoad
-      } else if (path == ".") {
-        (Nil, true)
-      } else {
-        val pathElements = path.split("\\.", -1).toList
-        val endsWithDot  = pathElements.lastOption.exists(_.isEmpty)
-        val realPathElements =
-          if (endsWithDot) pathElements.init
-          else pathElements
-        (realPathElements, endsWithDot)
-      }
-
-    def handleLoading(load: MorganeyLoading) =
-      load.modulePath map pathInformation getOrElse zeroLoad
-
-    def unapply(line: String): Option[(List[String], Boolean)] = {
-      val parseResult = LambdaParser.parseWith(line, _.loading).toOption
-      parseResult.map(handleLoading)
     }
   }
