+++ src/main/scala/me/rexim/morganey/autocompletion/extractors/LoadStatement.scala
@@ -1,0 +1,31 @@
+package me.rexim.morganey.autocompletion.extractors
+
+import me.rexim.morganey.ast._
+import me.rexim.morganey.syntax._
+import me.rexim.morganey.util._
+
+object LoadStatement {
+  val zeroLoad = (Nil, false)
+
+  def pathInformation(path: String) =
+    if (path.isEmpty) {
+      zeroLoad
+    } else if (path == ".") {
+      (Nil, true)
+    } else {
+      val pathElements = path.split("\\.", -1).toList
+      val endsWithDot  = pathElements.lastOption.exists(_.isEmpty)
+      val realPathElements =
+        if (endsWithDot) pathElements.init
+        else pathElements
+      (realPathElements, endsWithDot)
+    }
+
+  def handleLoading(load: MorganeyLoading) =
+    load.modulePath map pathInformation getOrElse zeroLoad
+
+  def unapply(line: String): Option[(List[String], Boolean)] = {
+    val parseResult = LambdaParser.parseWith(line, _.loading).toOption
+    parseResult.map(handleLoading)
+  }
+}
